requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
playwright
apscheduler
google-generativeai
python-dotenv
httpx
python-multipart
